Change u64 to u128

+++ src/child/src/store.rs
@@ -687,10 +687,10 @@ impl Store {
 
     // temporary put this here, should be in `ic_scalable_misc::helpers::token_canister_helper`
     pub async fn icrc_balance_of(canister: Principal, principal: Principal) -> u64 {
-        let call: Result<(u64,), _> =
+        let call: Result<(u128,), _> =
             api::call::call(canister, "icrc1_balance_of", (principal,)).await;
         match call {
-            Ok(response) => response.0,
+            Ok(response) => response.0 as u64,
             Err(_) => 0,
         }
     }
